Change naming in `createQuery` tests
